# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Project1 API",
    "description": "API for managing contacts",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/contacts/{id}": {
      "get": {
        "summary": "Get a contact by ID",
        "operationId": "getContactById",
        "tags": ["contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the contact",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single contact",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found"
          }
        }
      },
      "put": {
        "summary": "Update a contact by ID",
        "operationId": "updateContactById",
        "tags": ["contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the contact",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully"
          },
          "404": {
            "description": "Contact not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a contact by ID",
        "operationId": "deleteContactById",
        "tags": ["contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the contact",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully"
          },
          "404": {
            "description": "Contact not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "The unique identifier for the contact"
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the contact"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the contact"
          },
          "email": {
            "type": "string",
            "description": "The email of the contact"
          },
          "favoriteColor": {
            "type": "string",
            "description": "The favorite color of the contact"
          },
          "birthday": {
            "type": "string",
            "description": "The birthday of the contact"
          }
        },
        "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"]
      }
    }
  }
}